add fs

+++ server/server.js
@@ -21,20 +21,23 @@ app.post('/:year/:month/:day', (req,res) => {
     const path = "db/"+year+"/"+month+"/"+day;
     console.log("/"+year+"/"+month+"/"+day);
 
-    const data = {
+    const data = [{
         event_name: req.body.event_name,
         place_name: req.body.place_name,
         time_from: req.body.time_from,
         time_to: req.body.time_to,
         job_tag: req.body.job_tag,
         comment: req.body.comment
-    };
+    }];
 
     const content = JSON.stringify(data);
-
-    fs.mkdirSync(path, { recursive: true});
-    fs.writeFileSync(path+"/test.json", content);
-         
+    
+    if(!fs.existsSync(path)){
+      fs.mkdirSync(path, { recursive: true });
+      fs.writeFileSync(path+"/test.json", content);
+    }else{
+      fs.appendFileSync(path+"/test.json", "\n"+content);
+    }
 
     console.log(data);
     res.end();
